+++ NEWS.md
@@ -1,3 +1,5 @@
+- Added fishery catch data
+- updated fishery data to include licenses sold and license holder survey data
 - Added individual fish data
 
 # kootlake 0.2.7