+++ NEWS.md
@@ -1,3 +1,6 @@
+# kootlake 0.3.0
+
+- Deprecated `kl_estimate_na()`.
 - Added 2018 Gerrard peak count
 - Added fishery catch data
 - Updated fishery data to include licenses sold and license holder survey data